refactor: change class location in package (#37)
