--- api/src/main/kotlin/team/backend/domain/mapper/quiz/QuizMapper.kt
@@ -1,25 +1,0 @@
-package team.backend.domain.mapper.quiz
-
-import org.mapstruct.InjectionStrategy
-import org.mapstruct.Mapper
-import org.mapstruct.Mapping
-import org.mapstruct.ReportingPolicy
-import team.backend.domain.command.quiz.QuizCommand
-import team.backend.presentation.quiz.dto.QuizDto
-
-@Mapper(
-    componentModel = "spring",
-    injectionStrategy = InjectionStrategy.CONSTRUCTOR,
-    unmappedTargetPolicy = ReportingPolicy.ERROR
-)
-interface QuizMapper {
-
-    fun from(
-        request: QuizDto.AnswerRequest
-    ): QuizCommand.AnswerRequest
-
-    @Mapping(source = "correct", target = "isCorrect")
-    fun from(
-        request: QuizCommand.AnswerResponse
-    ): QuizDto.AnswerResponse
-}
