+++ api/src/main/kotlin/team/backend/domain/member/repository/MemberQuizAnswerStore.kt
@@ -1,0 +1,8 @@
+package team.backend.domain.member.repository
+
+import team.backend.domain.member.entity.MemberQuizAnswer
+
+interface MemberQuizAnswerStore {
+
+    suspend fun save(entity: MemberQuizAnswer)
+}
