+++ api/src/main/kotlin/team/backend/domain/member/repository/MemberReader.kt
@@ -1,0 +1,8 @@
+package team.backend.domain.member.repository
+
+import team.backend.domain.member.entity.Member
+
+interface MemberReader {
+
+    suspend fun get(id: Long): Member
+}
