+++ api/src/main/kotlin/team/backend/domain/member/repository/MemberStore.kt
@@ -1,0 +1,4 @@
+package team.backend.domain.member.repository
+
+interface MemberStore {
+}
