+++ api/src/main/kotlin/team/backend/domain/prompt/repository/PromptStore.kt
@@ -1,0 +1,4 @@
+package team.backend.domain.prompt.repository
+
+interface PromptStore {
+}
