+++ api/src/main/kotlin/team/backend/domain/quiz/command/QuizCommand.kt
@@ -1,0 +1,14 @@
+package team.backend.domain.quiz.command
+
+class QuizCommand {
+
+    data class AnswerRequest(
+        val quizId: Long,
+        val answer: String,
+    )
+
+    data class AnswerResponse(
+        val isCorrect: Boolean,
+        val answer: String,
+    )
+}
