+++ api/src/main/kotlin/team/backend/domain/quiz/repository/QuizReader.kt
@@ -1,0 +1,8 @@
+package team.backend.domain.quiz.repository
+
+import team.backend.domain.quiz.entity.Quiz
+
+interface QuizReader {
+
+    suspend fun get(id: Long): Quiz
+}
