+++ api/src/main/kotlin/team/backend/domain/quiz/repository/QuizStore.kt
@@ -1,0 +1,4 @@
+package team.backend.domain.quiz.repository
+
+interface QuizStore {
+}
