+++ api/src/main/kotlin/team/backend/domain/quiz/service/QuizService.kt
@@ -1,0 +1,8 @@
+package team.backend.domain.quiz.service
+
+import team.backend.domain.quiz.command.QuizCommand
+
+interface QuizService {
+
+    suspend fun answer(command: QuizCommand.AnswerRequest): QuizCommand.AnswerResponse
+}
