--- api/src/main/kotlin/team/backend/domain/repository/member/MemberReader.kt
@@ -1,8 +1,0 @@
-package team.backend.domain.repository.member
-
-import team.backend.domain.entity.member.Member
-
-interface MemberReader {
-
-    suspend fun get(id: Long): Member
-}
