--- api/src/main/kotlin/team/backend/domain/repository/member/MemberStore.kt
@@ -1,4 +1,0 @@
-package team.backend.domain.repository.member
-
-interface MemberStore {
-}
