--- api/src/main/kotlin/team/backend/domain/repository/quiz/QuizReader.kt
@@ -1,8 +1,0 @@
-package team.backend.domain.repository.quiz
-
-import team.backend.domain.entity.quiz.Quiz
-
-interface QuizReader {
-
-    suspend fun get(id: Long): Quiz
-}
