+++ api/src/main/kotlin/team/backend/presentation/quiz/QuizController.kt
@@ -4,8 +4,8 @@ import jakarta.validation.Valid
 import org.springframework.web.bind.annotation.PostMapping
 import org.springframework.web.bind.annotation.RequestBody
 import org.springframework.web.bind.annotation.RestController
-import team.backend.domain.mapper.quiz.QuizMapper
-import team.backend.infrastructure.service.quiz.QuizService
+import team.backend.domain.quiz.mapper.QuizMapper
+import team.backend.domain.quiz.service.QuizService
 import team.backend.presentation.quiz.dto.QuizDto
 import team.backend.response.SuccessResponse
 
@@ -19,6 +19,8 @@ class QuizController(
     @PostMapping("/v1/quiz/answer")
     suspend fun answer(@Valid @RequestBody request: QuizDto.AnswerRequest): SuccessResponse<QuizDto.AnswerResponse> {
         val answer = quizService.answer(quizMapper.from(request))
+        val sum = { x: Int, y: Int -> x + y }
+
         return SuccessResponse.from(quizMapper.from(answer))
     }
 }
